requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile
#
certifi==2019.6.16        # via requests, sentry-sdk
chardet==3.0.4            # via requests
envparse==0.2.0
idna==2.8                 # via requests
mattermostdriver==6.2.0
pymongo==3.9.0
python-logstash==0.4.6
pyyaml==5.1.2
requests==2.22.0
sentry-sdk==0.11.0
urllib3==1.25.3           # via requests, sentry-sdk
websockets==8.0.2         # via mattermostdriver
